requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.2
Django==3.2.7
djangorestframework==3.12.4
filelock==3.0.12
idna==3.2
importlib-metadata==4.8.1
kombu==5.1.0
platformdirs==2.3.0
prompt-toolkit==3.0.20
pytz==2021.1
ratelimit==2.2.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
zipp==3.5.0
